+++ UntitledBankApp/Services/AdminService.cs
@@ -27,7 +27,7 @@ public class AdminService
             return true;
         }
     }
-    public void SetCurrencyRate(CurrencyCode currencyCode, float rate)
+    public void SetCurrencyRate(CurrencyCode currencyCode, decimal rate)
     {
         if (_pseudoDb.Currencies.ContainsKey(currencyCode))
         {
