+++ UntitledBankApp/Services/ClientService.cs
@@ -150,7 +150,7 @@ public class ClientService
             if (foundAccount != null)
             {
                 account.Balance.Amount -= amount;
-                var convertedAmount =  ((float)amount * (1 / account.Balance.Currency.Rate)) * foundAccount.Balance.Currency.Rate;
+                var convertedAmount =  ((decimal)amount * (1 / account.Balance.Currency.Rate)) * foundAccount.Balance.Currency.Rate;
 
                 foundAccount.Balance.Amount += (decimal)convertedAmount;
